pyproject.toml
[project]
name = "dongzhou-tech-internship"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "ollama>=0.1.0",
    "python-dotenv>=1.2.2",
    "transitions>=0.9.3",
]
